[Memoria]: A?adido control para seleccionar la zona de memoria a visualizar. Corregido error en funci?n GetMem. Habilitado debug solo despu?s de seleccionar ROM

+++ src/Debugger.cpp
@@ -88,7 +88,7 @@ std::string Debugger::GetMem(WORD start, WORD end)
     end = (end & 0xFFF0)+0x000F;
     
     stringstream ss;
-    WORD row = start;
+    unsigned int row = start;
     while (row <= end)
     {
         ss << "0x";
